requirements.txt
django-toolbelt
pandas
Django
dj-database-url
dj-static
gunicorn
psycopg2
static
djangorestframework
markdown
django-filter
numpy
scikit-learn
matplotlib
django-picklefield